% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame containing edges with columns v1, v2, and w (weight).}

\item{init_node}{A numeric scalar representing the starting node.}
}
\value{
A numeric vector representing the shortest path to every other node.
}
\description{
Computes the shortest path from an initial node to all other nodes in a graph.
}
\examples{
wiki_graph <- data.frame(
  v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)
dijkstra(wiki_graph, 1)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
