% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar or integer.}

\item{b}{A numeric scalar or integer.}
}
\value{
A numeric scalar that is the greatest common divisor of the inputs.
}
\description{
Calculates the greatest common divisor (GCD) of two numbers using the Euclidean algorithm.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
