% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{France}
\alias{France}
\title{Shape files of France}
\format{
A data frame with 13 rows and 3 variables:
\describe{
\item{COUNTRY}{Name of country}
\item{REGION}{Name of regions}
\item{geometry}{lists of polygons}
}
}
\source{
\url{https://gadm.org/}
}
\usage{
France
}
\description{
A dataset containing geographic information of the polygons and the additional information
}
\keyword{datasets}
