\name{lab3package}
\alias{assignment3package}
\alias{lab3package}
\docType{package}
\title{
\packageTitle{Eucledian}
}
\description{
\packageDescription{Computes the greatest common divisor of two given integers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3package}
\packageIndices{lab3package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab3package}

Maintainer: \packageMaintainer{lab3package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
