% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3package.R
\docType{package}
\name{lab3package}
\alias{lab3package}
\alias{lab3package-package}
\title{Euclidean algorithm and Dijkstra's algorithm}
\description{
This package contains two functions. The first function uses the
    Euclidean algorithm to compute the greatest common divisor of two positive
    integers. The second function uses Dijkstra's algorithm to compute the
    shortest paths between nodes in a graph.
}
\section{Euclidean}{

See \code{\link{euclidean}}.
}

\section{Dijkstra}{

See \code{\link{dijkstra}}.
}

\author{
\strong{Maintainer}: Amer Memic \email{ameme698@student.liu.se}

Authors:
\itemize{
  \item Olof Josefsson \email{olojo524@student.liu.se}
  \item Shahreen Saleem \email{shahreen.saleem111@gmail.com}
}

}
