% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_5.R
\name{team_5}
\alias{team_5}
\title{Convert shape file into a data frame using team_5 lab2}
\usage{
team_5(file, tolerance)
}
\arguments{
\item{file}{is a file path to a shape file.}

\item{tolerance}{is the value used for thinning the polygon,default is 0.1.}
}
\value{
The return value is a data frame of the geographic information of
the polygons and the additional information.
}
\description{
This function is generated based on team 5's solution from lab 2.
team_5 function allows you to create a data frame from  a \code{geometry} variable.
}
\examples{
library("Lab3package")
data("Yemen")

df <- team_5(Yemen, 0.1)

\dontrun{
df \%>\%
  ggplot(aes(x = long, y = lat, group = group)) + geom_polygon(aes(fill =HASC_1 ))
}

}
