% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_6.R
\name{team_6}
\alias{team_6}
\title{Convert shape file into a data frame using team_6 lab2}
\usage{
team_6(file, tolerance = 0.1)
}
\arguments{
\item{file}{The file path to a shape file}

\item{tolerance}{The value used for thinning the polygon}
}
\value{
A data frame including geographic information of the polygons and the additional information
}
\description{
Convert shape file into a data frame using team_6 lab2
}
\examples{
library(Lab3package)
data("France")

df <- team_6(France, 0.1)

\dontrun{
library(ggplot2)
df \%>\%
  ggplot(aes(x = long, y = lat, group = group)) + geom_polygon()
}

}
