% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.r
\name{thin}
\alias{thin}
\title{Thin large shape files}
\usage{
thin(x, tolerance = 0.1)
}
\arguments{
\item{x}{is a shape file.}

\item{tolerance}{is the value used for thinning the shape file,default is 0.1.}
}
\value{
The return value is a data frame with the thinned geographical
information.
}
\description{
This function allows you to thin the number of points in shape files.
}
\examples{
library("Lab3package")
data("Yemen")
dframe <- thin(Yemen)

}
