% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia Graph}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Starting node}
  \item{v2}{Ending node}
  \item{w}{Edge weight}
}
}
\usage{
wiki_graph
}
\description{
A graph dataset containing edges and weights for Dijkstra's algorithm.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{datasets}
