% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra's shortest path algorithm
it calculates the shotest path fron initial node to every node in a graph}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a dataframe containing v1,v2,w}

\item{init_node}{the starting node}
}
\value{
vector conraining the shortest distance from initial node to every node
}
\description{
dijkstra's shortest path algorithm
it calculates the shotest path fron initial node to every node in a graph
}
\references{
https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)
}
