% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm implementation ...}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A graph}

\item{init_node}{a specific node in graph}
}
\value{
Vector containing smallest distances to every node in graph
}
\description{
Dijkstra algorithm implementation ...
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
