% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Shortest Path Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame with 3 columns: v1 (source), v2 (destination), and w (weight).}

\item{init_node}{A numeric scalar representing the starting node.}
}
\value{
A numeric vector containing the shortest distances from init_node to all other nodes.
}
\description{
Dijkstra Shortest Path Algorithm
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
