% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Finds the shortest path from initial node to every other node in the graph using Dijkstra's Algorithm.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph, init_node}{Takes in two arguments, the first one is a dataframe passed to graph argument and second one is the starting node to find the shortest distance from.}
}
\value{
Returns the shortest path to all the other nodes available from the initial node.
}
\description{
Finds the shortest path from initial node to every other node in the graph using Dijkstra's Algorithm.
}
\examples{
\dontrun{ 
wiki_graph <-data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph,3)
dijkstra(wiki_graph,1)
}
}
