% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame with the columns `v1`, `v2`, and `w`, where
`v1` and `v2` represent nodes and `w` represents the edge weight.}

\item{init_node}{A numeric scalar specifying the node where the algorithm starts.}
}
\value{
A numeric vector containing the shortest distances from
 `init_node` to all nodes in the graph.
}
\description{
Finds the shortest distance from an initial node to all other nodes
in a weighted graph.
}
\details{
The algorithm starts at the initial node, checks its neighbouring nodes,
and updates their shortest known distances. This process continues until
all nodes have been visited.
}
\references{
Wikipedia: Dijkstra's algorithm.
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
