% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\value{
The largest number that both numbers are divisible by.
}
\description{
Computes the greatest common divisor of two integers.
}
