% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidian algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
the greatest common divisor of a and b.
}
\description{
find the greatest common divisor of two numbers
}
\details{
The first algoritm to implement is the Euclidian algorithm. Assert that the arguments are numeric scalars or integers.
}
\examples{
euclidean(24, 36)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
