% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Finds the greatest common divisor.}
\usage{
euclidean(m, n)
}
\arguments{
\item{m, n}{It takes m,n as two integer numbers.}
}
\value{
It returns the greatest common divisor as integer.
}
\description{
Finds the greatest common divisor.
}
\examples{
euclidean(100,1000)
}
\references{
https://en.wikipedia.org/wiki/Euclidean algorithm
}
\author{
Vinod kumar Dasari
}
