% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean's algorithm
it finds the GCD of two numbers}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{the first number}

\item{y}{the second number}
}
\value{
the GCD of two numbers
}
\description{
euclidean's algorithm
it finds the GCD of two numbers
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm.
}
