% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm implementation ...}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The gcd between both numbers
}
\description{
Euclidean algorithm implementation ...
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
