% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Eucledian Algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{as numeric}

\item{y}{as numeric}
}
\value{
Greatest common divisor of x and y .
}
\description{
Implemantation of Eucledian Algorithm in R.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100,10000)
}
\references{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Wikipedia - Eucledian algorithm}
}
\author{
Marketos Damgios, Chrystoforos Spyretos
}
