% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanfunction.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Greatest Common Divisor}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Interger 1}

\item{b}{Interger 2}
}
\value{
Return Greatest Common Divisor
}
\description{
Calculate Greatest Common Divisor
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
