% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean function.R
\name{euclidean}
\alias{euclidean}
\title{Prints 'G.C.D'}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Interger 1}

\item{b}{Interger 2}
}
\value{
Return Greatest Common Divisor
}
\description{
Prints 'G.C.D'
}
