% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the GCD of two given integers using Euclidean Algorithm.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a, b}{Two input parameters to find GCD.}
}
\value{
Returns the GCD of given two inputs and the output is an integer.
}
\description{
Find the GCD of two given integers using Euclidean Algorithm.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
