% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Recursive Euclidean algorithm function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A positive scalar number}

\item{b}{A positive scalar number}
}
\value{
The maximum common divisor of the two numbers
}
\description{
Algorithm to find the greatest common divisor of two numbers
}
\examples{
euclidean(93164, 5826)
euclidean(1000, 250)

}
\references{
Euclidean algorithm - \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}

