% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric value.}

\item{b}{A numeric value.}
}
\value{
The greatest common divisor of a and b.
}
\description{
Finds the greatest common divisor of two numbers.
}
\details{
The function divides the numbers and saves the remainder.
It continues until the remainder is 0.
The last value of a is the greatest common divisor.
}
\references{
Wikipedia: Euclidean algorithm
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
