\name{lab3package-package}
\alias{lab3package-package}
\alias{lab3package}
\docType{package}
\title{
\packageTitle{lab3package}
}
\description{
\packageDescription{lab3package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3package}
\packageIndices{lab3package}

}
\author{
\packageAuthor{lab3package}

Maintainer: \packageMaintainer{lab3package}
}

\keyword{ package }
