\name{lab3package-package}
\alias{lab3package-package}
\alias{lab3package}
\docType{package}
\title{
\packageTitle{lab3package}
}
\description{
\packageDescription{
  The package includes functions for calculations using the euclidean algorithm and the
dijkstras algorithm ...
}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3package}
\packageIndices{lab3package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Johan Marbinah}

Maintainer: \packageMaintainer{author}
}
\references{
~~ Literature or other references for background information ~~
https://en.wikipedia.org/wiki/Euclidean_algorithm 
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
