\name{Lab3package-package}
\alias{Lab3package-package}
\alias{Lab3package}
\docType{package}
\title{Lab3package
\packageTitle{Lab3package}
}
\description{
\packageDescription{Lab3package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab3package}
\packageIndices{Lab3package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab3package}

Maintainer: \packageMaintainer{Lab3package}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{


euclidean(53253, 54)
}
