% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki graph
a graph used to test dijkstra's algorithm}
\format{
a data frame with 18 rows and 3 variables
\describe{
  \item{v1}{starting vertex}
  \item{v2}{ending vertex}
  \item{w}{weight}}
}
\usage{
wiki_graph
}
\description{
wiki graph
a graph used to test dijkstra's algorithm
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\keyword{datasets}
