% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki graph data}
\format{
A data frame, 18 x 3
\describe{
  \item{v1}{from node}
  \item{v2}{to node}
  \item{w}{weight for edge between nodes}
}
}
\source{

}
\usage{
wiki_graph
}
\description{
Wiki graph data
}
\keyword{datasets}
