% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph Data}
\format{
A data frame with 18 rows and 3 variables.
\describe{
  \item{v1}{index of the source edge}
  \item{v2}{index of the destination edge}
  \item{w}{weight of the path}
}
}
\source{
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Wikipedia - Dijkstra's algorithm}
}
\usage{
wiki_graph
}
\description{
An example data frame with 18 rows and 3 variables for the Dijkstra function. Based on the example from the \href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Wikipedia - Dijkstra's algorithm page}.
}
\keyword{datasets}
