% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Sample graph for Dijkstra algorithm}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Edge start node}
  \item{v2}{Edge end node}
  \item{w}{Edge weight}
}
}
\source{
Wikipedia Dijkstra Algorithm
}
\usage{
wiki_graph
}
\description{
A dataset containing nodes and edge weights based on Wikipedia.
}
\keyword{datasets}
