% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example graph information}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{vector of nodes where edges start}
  \item{v2}{vector of nodes where edges end}
  \item{w}{weight of the edges}
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstras_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the nodes and edges of an example graph
}
\keyword{datasets}

