% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yemen}
\alias{Yemen}
\title{Data description}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 11 columns.
}
\source{
\url{https://gadm.org/}
}
\usage{
Yemen
}
\description{
"Yemen" is created by read_sf("gadm36_YEM_1.shp") file form online source.
}
\keyword{datasets}
